Changes to PCoA files and ASV filtering
When providing an ASV list to remove or analyze, the PCoA plots were not being referred to the appropriate file, and so were not being properly filtered. Now when you provide an ASV list and select "remove" or "analyze", the associated ASVs will be removed and represented properly in the PCoA plots. Oops.