Changes to metadata file and all plots
Made changes to metadata file structuring so that the dropdown menus are populated in alphabetical order.

Changed the colour of facets to white from grey.

Restructured the way the bubble plot subsets specific taxa (now just checks to see if there is something entered in the box) I also added a new "remove samples" option where the user can type in specific taxonomies to be removed from the plot without affecting abundance calculations.