Minor fixes and reorganizing
Changed the fill colour for the read plots so they weren't needlessly colourful.

Changed how representative sequences were added back into the long-form data tables.

Added a feature that allows users to include all samples in the bubble plot, regardless of whether there is ASV data. This is pretty edge-case, and is really meant for when people (me) want to show a subset of taxa that may or may not be present in a sample. Normally, ggplot will not include these (since there is no y-data to plot). To sloppily get around this, I included an option to add a fake taxon after all abundance calculations. Thus, so long as this taxon is included in the subset, all samples will be plotted. A user (me) can then remove this fake taxon (which belongs to a fake phylum) from the image and preserve everything else.

If anyone has a suggestion for how to do this better, let me know.